Return the presign expiry and close the slice-2 review findings
Against the PM brief this slice owed an expiry in the response, a
performance pin, and the deletion of the slice-1 smoke test. It also
carried findings from a read of the whole path.

The presign's expiry is now returned rather than computed and discarded.
S3DocumentStorage answers with a PresignedUpload carrying the presigner's
own expiration, so what the caller is told is what the signature encodes
rather than a number recomputed from the requested duration. A client can
now fail fast instead of discovering a spent URL mid-upload.

Product limits moved into DocumentUploadRules and the cap rose from 5MB to
25MB - the brief asks for tens of MB, and a 300dpi scan of a 40-page past
paper lands around 15-20MB. The expiry is now passed into the storage call
rather than held there, so infrastructure carries no product knowledge and
the two numbers a reader looks for sit together.

presignUpload takes a PresignedUploadRequest that guards its own values.
This reverses an earlier judgement: I argued storage sat inside a trusted
region and needed no guard, but that region is a property of today's caller
graph, not of a public method on a singleton. Three of its four inputs -
blank key, impossible length, spent expiry - sign successfully and hand back
a plausible URL that can never be used, which is exactly the failure a guard
is for. The rule this settles on: public methods on injectable components
guard their inputs; values already inside a validated record do not.

An absent X-Admin-Key now answers 401 rather than 400. Binding the header as
required meant Micronaut rejected it before the validator ran, so a caller
could tell a missing key from a wrong one - a free oracle on an admin route.

Tests no longer import application constants. A test asserting against
MAX_CONTENT_LENGTH imported from production moves in lockstep with it, so
the value was untested by construction; restating it test-side made raising
the cap fail three tests that would otherwise have passed silently. SVG left
the upload allow-list: it is XML that can carry script, and the account that
uploads documents is the worst one to expose to stored XSS.

Also: save()'s return value is used rather than relying on Hibernate
mutating the argument; a StringUtil replaces three different blank-check
mechanisms; ErrorResponseMapper gains its own test; content length is int
throughout instead of mixing Integer and long.

AdminDocumentPostPT pins the create at 1 statement, confirming the
@Generated read-back rides Postgres's insert-returning. All three tiers
green: test 162, integrationTest 110, performanceTest 11.

+++ src/test/java/com/practiq/service/document/StagedDocumentUploadTest.java
@@ -14,7 +14,7 @@ class StagedDocumentUploadTest {
         String key = "cf5c1c1e-8a7d-4b0e-9f2a-6d3b7c8e1f04.pdf";
         String filename = "physics-notes.pdf";
         String sourceSpec = "AQA GCSE Physics";
-        long contentLength = 2048;
+        int contentLength = 2048;
 
         StagedDocumentUpload staged =
                 new StagedDocumentUpload(key, filename, sourceSpec, MediaType.APPLICATION_PDF_TYPE, contentLength);
